requirements.txt
numpy
pandas
djangorestframework
